parser: several type simplifications
Remove Named types. In a coming update they will be gone. The only
kind of type declarations will be struct (maybe renamed to class)
and interface. The notion of Underlying goes away with it.

Take parameter names off the function type. They are not part of the
type, because func min(x, y int64) is satisfied by the type signature
func min(a, b int64).