parser: structs are now classes
This is a big simplification of the type system. Only two kinds of
types can be declared now: classes and interfaces. A class has a
bunch of methods declared on it (making it REPL-compatible). It is
equivalent to a Go struct with pointer receivers. Interfaces work
just as in Go.

This type system diverges from Go in several ways now, all of which
are limiting and a bit frustrating. Naming was really useful and I'm
sad to see it gone. However, it could come back later. The main
reason for doing this is it makes it easier to play around with other
type changes, and makes for a simpler frontend for a 1-person
language project.

(It will be great to do away with the concept of untyped bools and
untyped strings in the type checker.)

Still potentially completely compatible with Go objects by treating
named types as opaque.