parser: prep for $$
To encourage me to use ng in my day-to-day life I'm embedding a shell
inside the REPL. (Yes, it's completely backwards and really weird.
That's part of the appeal.)

If I get it working well enough that I can invoke cd and vi, I can
replace bash in my life with ng, which will push to me to hurry up
and get basic control flow working.