parser: disable broken table index logic
A lot of the TableIndex logic depended on the incorrect notion that a
column name string would only be passed in as a string literal. In
the statement t[x,y], x could be a string or an int.

For now, disable so I can get on with map index support and revisit
soon. It's not clear if the syntax is broken, or the parser just needs
a slightly more general AST construction.