eval: rewrite the interpreter
The interpreter was rewritten in this commit to adopt two
fundamental design changes.

Firstly, objects are now Go objects with the same type shape.
This makes calling Go code and sharing data with Go code much
simpler, and we can ditch a lot of impossibly complex booking
the interpreter was doing. The downside is we are now limited
by the power of the reflect package: for example the inability
to add methods or create interfaces (as of Go 1.7) means we
cannot do some things. But I believe it is easier to fix Go
than it is to get the old interpreter's design correct.

Secondly, instead of propagating interpeter bugs as errors we
now use an internal panic object. This is makes the code cleaner,
and is important for interaction with Go code. For example, if
an interpreter bug appears in the Neugram function called by a Go
function, it needs to be translated into a panic to get back out
to the original Neugram interpreter.

Lots of TODOs and broken tests, but this is a nice checkpoint.