eval: basic method support
This commit introduces a "method pool". It is a set of pre-declared
Go types for certain method signatures, that can be used to
construct neugram types with methods of the same signature.

It is a temporary hack. Long-term, it will be replaced by either a
Go source backend that can be comppiled to a plugin and loaded, or
use of the non-existent reflect.NamedOf.

On the Go 1.9 timeline, I expect plugins to happen first.