typecheck,eval: implement dup-decl-check
This CL implements a declaration duplicates check.
There is a slight departure from proper Go validation, under the more
dynamical nature of ng: an identifier is labelled as a duplicate if
it redeclares another pre-existing variable.
For example:
  import "fmt"
  fmt := 42
is not considered to be redeclaring `fmt`.

Add tests that detects duplicate declaration of variables.
Updates tests for the more stringent declaration checks.

Fixes neugram/ng#30.