eval, typecheck: do not pollute scope on import
We typecheck .ng files twice. Once when we import the file,
and one again line-by-line as we evaluate the file.

The evaluator was not loading a new typecheck scope for the
line-by-line evaluator, meaning top-level declarations in an
import were polluting the typecheker's scope for the importing
program. Fix this by pushing an ephemeral typecheck scope.

Maybe we shouldn't typecheck .ng imports twice.