typecheck: protect type info with a mutex
The typechecker can be used by multiple goroutines simultaneously. That
means we need a way for the evaluator to create new empty scopes and
make it safe for concurrent use. This commit introduces NewScope and
hides type access behind a mutex for this purpose.

Some future thought should be put into the type checker modifying the
syntax tree. It has generally been fair game for typecheck to annotate
types, for example ElideError. But in a concurrent world, we need to
make sure the annotation is only applied once, and is definitely
applied before concurrent reads. We may already be there, but it
deserves some time reading the code.

Updates #37