eval: avoid value boxing call on method recveiver
The method recveiver can end up in an inadvertent data race if
reflect.Value.Interface is used to box the value while other
goroutines are attempting to call (perfectly valid) methods on
the recveiver simultaneously.

Avoid the problem by not doing it. This fixes the data race that
eval/testdata/sync.ng was showing under the race detector.

Fixes #37.