ngcore: interface for the neugram interpreter
The ngcore package is intended to tie the parser and evaluator
together into a complete interpreter, behind an interface useful
both to the ng command line binary and programs that want to
embed neugram.

There is much more to do, but this is enough to get a very
basic jupyter kernel working.

Updates #16