typecheck: move argument formatting into errorfmt
Several typechecker errors were failing to format expressions.
Turns out, the style of formatting we want for all of these
errors is the same, so move the logic into the error creation
function.

(This will make it easier to tune format.Type later on,
right now it is too noisy.)

Also rename it from errorf to errorfmt to keep vet happy.
I'm not happy about this, vet shouldn't be breaking on my perfectly
fine program. But it's faster to fix it here for now.