gengo: basic hoisting of types with methods
There is a lot going on here. This is the first step in a body of
work around method handling.

It turns out a methodik cannot generally refer to any in-scope
variable, without constructing a type that we cannot convert to Go.
(A future CL will add checks to the type checker to make sure such
types aren't constructed.) Maybe this is worth a blog post.

When generating Go, we hoist method definitions to the global scope.
This requires hoisting all variables and types that a method can
depend on out of the primary init function. So the restriction we
are applying now is: to refer to a free variable in a methodik it
must be possible to hoist that variable to the top-level.