eval, gengo: replace method pools with plugins
There is no way (presently) to use run time reflection in Go to
create types with new methods. The evaluator contained a hack to
use some pre-defined types with methods to emulate defining new
types, but this supported only an extremely limited set of
pre-defined method signatures.

Instead, we move to generating a plugin that defines a new
type with methods. These methods contain the minimum reflect-based
trampoline necessary to return control to the evaluator.

Still missing support for several features such as pointer
recievers and methods on non-struct types, but those are now
possible to implement.