eval: get method receiver values working
Remove the use of reflect.StructOf.
Because StructOf does not yet generate wrapper methods, types
wrapped with it do not satisfy interfaces, which defeats a lot
of the value of methodik definitions.

This means we need a definition of all field types that is static
at the point the plugin is defined. For types that have been
named in the running evaluator, we expand them out into
unnamed underlying types. This works because the types defined in
the evaluator have been created using the reflect package, which
makes them equivalent to unnamed static types.

Future work: One place where this is going to break down is when
the field of a methodik has the type of another methodik defined in
the evaluator scope. We will need to use the named type directly.
To do that we need to move from using simple main.go plugins to
building up a collection of packages that are shared by subsequent
plugins, so one can use a type defined in another.